Fix mobile build errors
