+++ zcncore/transaction.go
@@ -656,107 +656,6 @@ func (t *Transaction) isTransactionExpired(lfbCreationTime, currentTime int64) b
 	sys.Sleep(defaultWaitSeconds)
 	return false
 }
-func (t *Transaction) Verify() error {
-	if t.txnHash == "" && t.txnStatus == StatusUnknown {
-		return errors.New("", "invalid transaction. cannot be verified.")
-	}
-	if t.txnHash == "" && t.txnStatus == StatusSuccess {
-		h := t.GetTransactionHash()
-		if h == "" {
-			transaction.Cache.Evict(t.txn.ClientID)
-			return errors.New("", "invalid transaction. cannot be verified.")
-		}
-	}
-	// If transaction is verify only start from current time
-	if t.txn.CreationDate == 0 {
-		t.txn.CreationDate = int64(common.Now())
-	}
-
-	tq, err := NewTransactionQuery(_config.chain.Sharders)
-	if err != nil {
-		Logger.Error(err)
-		return err
-	}
-
-	go func() {
-
-		for {
-
-			tq.Reset()
-			// Get transaction confirmationBlock from a random sharder
-			confirmBlockHeader, confirmationBlock, lfbBlockHeader, err := tq.getFastConfirmation(context.TODO(), t.txnHash)
-
-			if err != nil {
-				now := int64(common.Now())
-
-				// maybe it is a network or server error
-				if lfbBlockHeader == nil {
-					Logger.Info(err, " now: ", now)
-				} else {
-					Logger.Info(err, " now: ", now, ", LFB creation time:", lfbBlockHeader.CreationDate)
-				}
-
-				// transaction is done or expired. it means random sharder might be outdated, try to query it from s/S sharders to confirm it
-				if util.MaxInt64(lfbBlockHeader.getCreationDate(now), now) >= (t.txn.CreationDate + int64(defaultTxnExpirationSeconds)) {
-					Logger.Info("falling back to ", getMinShardersVerify(), " of ", len(_config.chain.Sharders), " Sharders")
-					confirmBlockHeader, confirmationBlock, lfbBlockHeader, err = tq.getConsensusConfirmation(context.TODO(), getMinShardersVerify(), t.txnHash)
-				}
-
-				// txn not found in fast confirmation/consensus confirmation
-				if err != nil {
-
-					if lfbBlockHeader == nil {
-						// no any valid lfb on all sharders. maybe they are network/server errors. try it again
-						continue
-					}
-
-					// it is expired
-					if t.isTransactionExpired(lfbBlockHeader.getCreationDate(now), now) {
-						t.completeVerify(StatusError, "", errors.New("", `{"error": "verify transaction failed"}`))
-						return
-					}
-					continue
-				}
-
-			}
-
-			valid := validateChain(confirmBlockHeader)
-			if valid {
-				output, err := json.Marshal(confirmationBlock)
-				if err != nil {
-					t.completeVerify(StatusError, "", errors.New("", `{"error": "transaction confirmation json marshal error"`))
-					return
-				}
-				confJson := confirmationBlock["confirmation"]
-
-				var conf map[string]json.RawMessage
-				if err := json.Unmarshal(confJson, &conf); err != nil {
-					return
-				}
-				txnJson := conf["txn"]
-
-				var tr map[string]json.RawMessage
-				if err := json.Unmarshal(txnJson, &tr); err != nil {
-					return
-				}
-
-				txStatus := tr["transaction_status"]
-				switch string(txStatus) {
-				case "1":
-					t.completeVerifyWithConStatus(StatusSuccess, Success, string(output), nil)
-				case "2":
-					txOutput := tr["transaction_output"]
-					t.completeVerifyWithConStatus(StatusSuccess, ChargeableError, string(txOutput), nil)
-				default:
-					t.completeVerify(StatusError, string(output), nil)
-				}
-				return
-			}
-		}
-	}()
-	return nil
-}
-
 func (t *Transaction) GetVerifyOutput() string {
 	if t.verifyStatus == StatusSuccess {
 		return t.verifyOut
